Deactivate "change accounts" menu without hiding the action bar logo
The app has a config option called `showChangeAccountsUi`, which is used
to enable/disable the logo in the action bar, which is used as a button
to open the "change accounts" menu.

In E-kirjasto, we want to disable that menu, but still show the logo. To
do this, a new build config option `showActionBarLogo` was added (this
new option is only in effect if `showChangeAccountsUi` is false).

If `showChangeAccountsUi` is false and `showActionBarLogo` is true, the
logo is shown, but does not act as a button. These settings are now the
default in E-kirjasto.

This code could even be contributed into The Palace Project, but no
plans to do that for now.

Issue: SIMPLYE-196