Settings links initial feedback link
This is the first hardcoded feedback link that worked. There is an
existing bug that prevents links from being displayed so for now due to
deadline pressures the links will be hardcoded.

In the future the links should be softcoded as the backend already seems
to be showing them properly.