Add feedback, accessibility statement and privacy policy links
This is the "translation hack" solution which causes less potential issues if
things are changed upstream. Some of the semantic meaning of the links is
broken in favor of avoiding future conflicts. For that reason there are plenty
of comments to relatively little code. There was a previous solution where the
links were done by adding more fields that were properly named, but that
approach resulted in rather serious potential merge conflicts as well as almost
100 lines of code per link.

An alternative solution is possible where overriding the original classes is
used instead. This approach is more in line and a stepping stone to that.

Refs EKIR-45, SIMPLYE-217