Change constraints in book_detail
Added a constraint between top of buttons and bottom of author, because
the buttons would cover the author when book had a long name. Removed
constraint of bottom of buttons to bottom of cover image, because it made
the buttons be on top of text or make the buttons dissapear when text was
scaled or name of a book was long. Removed a plus from @id call because
it was not needed.

Ref EKIR-95