Change settings to use theme and be scrollable
Added the style to all the buttons and texts, as
well as the background. Added a scroll view because
the page would be cut off on smaller screens. Added
the logo and deleted the old creation method of the
logo to match the other views. Added the changes made
in ekirjasto-themes.This page is also working in
dark mode.

To note: Settings and other text at the top of the
screen is not centered. Buttons are width of the screen
so on bigger screens they are too wide.

Ref SIMPLYE-317 SIMPLYE-324