Set missing link and delete about app button
Set the link for the user agreement webpage in
EkirjastoDocumentStoreConfiguration, and deleted
the aboutApp button from layout file account_ekirjasto.xml,
and from EkirjastoAccountFragment. Moved the feedback button
to the spot aboutApp was deleted from, and adjusted the
space between the top and bottom to count for the missing
button.

Ref SIMPLYE-334