Fix crash when pressing "Back to service" during Suomi.fi login
The code assumed that the response from the login was a token, but the
response is completely empty if the user cancels the login.

Fixing the crash made canceling the login act like an error, but it's
not an error, so I added a "cancel" event to the Suomi.fi login flow,
so that canceling is handled a bit more gracefully.

Issue: SIMPLYE-349