Fix crash when skipping back in audiobook player
The app crashes when skipping back 30 seconds in the audiobook player,
if the player is less then 30 seconds from the start (if trying to skip
to before the beginning).

The crash is in the audiobook player library, which is not in this
repo. Updating to the next minor version fixes the crash, but that
version also updates lots of libraries, and updating those libraries
required updating classes unrelated to the audiobook player.

Still work in progress, because I need to clean up commented out code,
and check that I didn't break anything in the PDF reader or other
modules that use the updated libraries. And I need to rebase these
changes so that it's not based on the digital magazines branch...

Issue: SIMPLYE-333