Allow cleartext network traffic for localhost (127.0.0.1)
Opening books broke because a recent commit disabled allowing cleartext
network traffic (HTTP instead of HTTPS). The change was made because of
a security warning in Google Play Console. However, this broke loading
book resources from localhost (127.0.0.1).

This commit changes the config to allow cleartext network traffic
again, but this time only for localhost. This fixes opening books,
hopefully the security warning won't come back. Even if it does, this
isn't really a security issue. If the warning comes back it might be
possible to use `file://` instead of `http://127.0.0.1/`.