Implement in-app language options
Created a new file LocaleHelper to add the
chosen locale to memory. Override attachBaseContext
in MainApplication and MainActivity, so the new language
is used, set the default language as finnish. Change the
language lookup for the links and feed to use the chosen language.
Add buttons (unstyled) from which the language can be changed.
Added a popup telling about restart. When accepted, app restarts
and language changes.

TO NOTE: This messes up the dark/light modes, not updating to
the changed one until app restart.Problem is the override of
attachBaseContext in mainApplication.

REF ekir-185