Rewrite CI workflows, add Fastlane uploads
This PR rewrites and fixes the CI workflows (in GitHub Actions) and adds
automated build uploads to Google Play Console using Fastlane. There's
still some work left here (the release workflow and documentation), but
this should be a good starting point (I'll finish the rest this week).

The documentation is also updated, but there's definitely still some
work left there, I'll make another PR soon for that (I'd like to keep
all documentation in the repository).

The main change for developers is that this commit removes the use of
`simplified-app-ekirjasto/src/main/assets/secrets.conf`, which was used
to supply the Transifex token. This is now moved to local.properties, so
all build and runtime secrets are defined through there, making life a
bit nicer for developers.

Issue: SIMPLYE-380
Issue: SIMPLYE-227