Update CI workflows and documentation; Write release checks
This commit also fixes some lingering issues with Transifex
localizations. There's still a "ghost string" with the ID 'app_name',
which isn't defined or used anywhere in the repository, but it keeps
coming back on every Transifex download... The string ID was changed,
and the new string is correctly translated, so the old string can be
safely ignored.

Issue: SIMPLYE-380