Obfuscate secrets
It's impossible to completely hide secrets in a client application, but
at least we can make it a bit more difficult to see them.

Palace had them in an asset file, so it was enough to just extract the
APK, and you could see all secrets handily in one file. With the way
secrets are stored here, you would need to decompile the code, find
strings, and decode those into the secrets.

This makes it a bit more difficult, but still not impossible (it's
impossible to make it impossible, if the secrets are stored
client-side). The current method could definitely still be improved, I
have a few ideas on how to do that.