Fix dark mode and link languages
Changed LocaleHelper to return the base context
it is given, as we restart when language changes
so making a new context is not needed. This fixed dark
mode. Changing color mode reset Locale information, but
correct information is found always in LocaleHelper, so
changed LanguageUtil.kt, MagazinesFragment.kt and
CustomHTTPInterceptors.kt to get their language from
LocaleHelper.

TO NOTE: If device language is switched during app running,
it resets some of the in-app languages to system language
until restart.

Ref EKIR-185