Remove unused .ci submodule and old CI scripts
The .ci submodule is from Palace, and it's not used for anything here.
We also *can't* use it, since it's a Palace repository, so we can't
modify the scripts to make them work for E-kirjasto.

I also removed other unused/outdated CI scripts, they're all now under
the scripts directory (and they're not just for CI, but can also be run
locally). Having the old scripts there can cause confusion, since you
don't know which scripts to use, and if the script is broken or if
there's something wrong with your environment, etc., so it's best to
just delete them. We can always refer to the old scripts in Git history
or in the Palace repositories, if we need to.