Update version code calculation
We want to upload multiple flavors of the same build to Google Play
Console, but Google Play Console requires that the version code is
increased for every upload, meaning that the same (or lower) version
code cannot be uploaded again. This in turn means that we would have to
perform another full build before uploading another flavor.

This commit changes the version code calculation. It's still time-based
(epoch timestamp, but starting at 2021-03-15 09:20:00 UTC), but the
last digit zeroed, and then each flavor changes the last digit to:
- production: 1
- beta: 2
- dev: 3
- ellibs: 4

This way, it's possible to upload all flavors without rebuilding,
although the flavors have to be uploaded in the above order.

Also update the ekirjasto-theme submodule to the latest commit.