Make book page more accessible
Deleted the word button from content
descriptors and changed some words to more
descriptive ones in stringsCatalog.xml. Added
containers around the description and details
in book_detail.xml so that the title and content
are read together, this also makes it so, that all
details are read in one passing. Also made the
more button as wide as the whole screen, instead of
just clickable in the corner. The title and author
are now read in one go when hovering the area, although
the name and author are read twice: once in the cover
image description and once as the seen texts.
In fragment_tutorial.xml made the tab buttons
background always white so the circles show in dark mode.

Ref EKIR-253, EKIR-255