Minor CI fixes
In CI workflows, use `!cancelled()` instead of `always()`, so that the
step only runs if another step failed, not if the whole workflow has
been cancelled.

In Ruby, `string.strip!` will return the string only if there was
whitespace to strip, so the old code might have sometimes returned
`nil`.