Adjust recycle view layout for scaling
The type text on feedWithoutGroups would
disappear under the buttons when text size
was increased. So, constrained the buttons to
the bottom of the type text and made the cell
height wrap_content, so that everything was always
visible. The size still stays consistent because the
cover image is normally the constraint for the height
of the cell, and it is consistent. This case only changes
when the size of the text increases too much.
Also reverted some changes in book_detail, as the changes
affected negatively the layout.

Refs EKIR-253, EKIR-257