Add basis for post and get functionalities
Changed the event to include also the ekirjasto
token. This token might be available in some other way
so that it's more fresh. Current version does not work
and all calls return error 401.
For both lookup and post, added a task and request, and
put the functionalities into profiles controller, which
might not be the correct place for them, but this way I
managed to get the calls to go through on non-main thread.
Some files had only some prints added as to observe the
ekirjasto token ,which currently by our guess should be the one
to use as auth for the lookup and post, but currently does not
work.