Add a working scaling implementation
Added a new button into settings page that
gives a pop up with text scaling options.
Choosing one triggers a MainActivity reload
that scales the text. For this to work, added into
AccountFragment another listener that posts new
events called TextSizeEvent. This is listened then
by MainActivity that handles the changes. The information
is saved into AppCache where it is pulled every restart.

TO NOTE: Not compatible with language choice code
as they both work with attachBaseContext, and changing
the context messes with the transifex.

Ref EKIR-251