Fix crash from back button in Suomi.fi login
Application crashed when pressing the "Back to service"
button in Suomi.fi login. This was fixed earlier, but now
emerged again. This checks if the content contains the
error message that shows on response, and if it does, the
login just gets cancelled. Before we only checked if the
content was empty.