Canceling passkey only closes the prompt
In ViewModel, added a description and log for the
GetCredentialCancellationException, aka. for when
user cancels the passkey login. Also added a boolean
which is set to true when the error is a cancellation.
THis value is checked in PasskeyFragment, and if this
is the cause of the error, we only cancel the login, and
don't trigger failure actions.

Ref EKIR-260