Add new preferences fragment, move functionalities
Added a new PreferencesFragment, in which user can
find language and fontsize options. Added a switch that
doesn't have a real functionality, other than enable and disable the
buttons. Moved the language and font options to the new fragment.

Chaged the font choice so that language choice does not break and font
size is updated without needing to restart the application.

Added a util class that handles the shared preference calls for font
size. This should probably be moved to the appCache class.

Ref EKIR-251