Add a new behavior triggered by preferences
Added toolbar config into preferences, so the
linking works when the up arrow is pressed. The
behaviour is set as pop stack, unless stack empty,
when pop to accounts view. THis makes it so that when
mainActivity is reloaded and the stack is empty, it
still links to accounts fragment.