Add dependents fragment
This pr adds the dependents fragment and all
its functionalities. The fragment creation is
triggered on the settings page, with a button
placed under passkey creation. The button is
only shown when logged in.

The fragment consists of a button to get the dependents,
an info box to present different information to the user,
email field to enter the email to, and a button to post it,
as well as a button to return to add another dependents.

The get and post calls are handled in the
DependentsViewModel, and the fragment observes both error
and dependents data. When getting dependents, a call to
the circulation to get the fresh ekirjastoToken is made
and then to the backend to get the dependents.

The post is made by adding the dependent in the
body of the request. We assume the child's language
should be the same as the parent's.

From the circulation login we now extract
patronpermanentID along with the circulation token.

This pr also adds the handling of two new
authentication_document links. One is for the dependent
lookup, other for the post. They are added into
the AccountProvidersJSON,
AccountAuthenticationCredentialsJSON and
AccountProviderAuthenticationDescription.

NOTE: As said in Magazine implementation, changing
the authentication_document causes the one in memory
to become stale. Error on boot happens because of it,
but only once. If the app is restarted, this is fixed,
but this happens for all users with old cache.
This is helped by clearing app data, but a better
handling needs to be done.

REF: EKIR-235