Fix crash from stale accountJSON
If we add URIs that are stored in the accountJSON, the
first time we open the app after the update, the information in
the file is stale and the new URIs can't be found there. This
would cause the boot to fail, and the account to be recreated on next
launch, fixing the issue. This update fixes the situation by
returning "empty" url if the links are not yet in the file. This file is
updated right after boot, so the epmty URIs are overwritten before the
user even does anything.

Also created a better function description.