Add catalog reload to login
Usually login triggers a catalog reload, but it
is not triggered from the viewmodel. So sometimes
the loans/reservations view shows empty despite there
being books on loan/reserve. The catalog feed reload
is done now on login. It is triggered 1-3 times during the
login (for some reason) but it only gets the remote feed
once and the other times are local feed, so the remote is not
called more times than before.

REF: EKIR-338