Add audiobook manifest on book load
Audiobook manifest was saved only on book
open, making closing the app or logging out
before opening the book make the book unusable.
This is now fixed by adding the manifest file to
memory when the book is loaded. The manifest file
is extracted from the zip file and added to local
database.

The errors that audiobook player was giving were
nondiscriptive, so added some new strings to help
with recognizing where the errors are happening.

Also added some helpful logs.