Add an information popup when not enough space
Added a popup that is shown when a book that is downloading
is too big to fit on the device. This popup is shown both
in single book view as well as the catalog. The popup also
interrupts the download. The popup informs user about the
remaining free space on their device, as well as how much
more space they need to free to be able to load the book.

Canceling a download causes the book to be removed from the
local database, and thus the loans page would incorrectly
not show the book as one of your loans. This is now fixed by syncing
the account with the backend when a download is cancelled (either
by the user or the file being too big). The sync gets the user's
loan information from the backend and thus updates the local
database with correct information.

Also added a onClick listener to the loading cell in the catalog
view, so that user can navigate to the bookDetail view even when the
book is loading.

Deleted the creation of an empty button space left of the load
button, that is useless and places the load button to
incorrect spot.

REF EKIR-326