Change the visuals of catalog cells
Increasing the size of the text on the device
caused text to show too little on some of the
buttons. The loan information was also very
small placed on a button and did not adjust in size.

The buttons are now located on their own row below the
cover image.

The loan time is now placed on a text area under the book
meta information in a yellow box. The box remains on the bottom of the
view close to the buttons, close to where the old buttons were.

On reservations, the time till the book is available is shown in
the same yellow box. The time is shown in days, and if under 2
days, in hours.

In order to help the views on smaller screens and bigger text sizes, the
book's name and authors can now extend to two rows before being
ellipsized.

Ref: EKIR-375, EKIR-363, EKIR-377