Pressing loan button opens the login page
Pressing loan button when not logged in used
to throw  the user into the settings page.

This change makes it so that the user is shown
the login page when the loan/reserve button is
clicked when not logged in.

When the fragment is showing, we hide the bottom
toolbar, so the view looks as close to possible
to the login screen you see when you open the
application. The only difference is the green
line shown on the bottom of the screen,
To ensure that the toolbar buttons are shown
correctly when pressing the back button and closing
the login screen, we set pressing the back button in
the main fragment to make the buttons visible.

TONOTE: Logging in empties our backstack and aways
shows the catalog when the login is finished. Thus
after loging in, the user isn't returned to the
book they started the login from, and pressing the
back button in that situation causes the app to return
to the basic catalog view (if we are in a particular lane)
or close the app (if in the main catalog view). This
behaviour is not different from the original behaviour,
but should be rectified for better user experience,

REF: EKIR-366, EKIRJASTO-70