Change MainApplication instance init spot
There was crashes that were seemingly caused by
boot fragment calling the mainApplication instance
before the value for the INSTANCE was set, so
this pr moves the initialisation of the instance
to the earlies spot in the onCreate.

As we were unable to recreate the crash, and the crash
was most likely invisible to the user, there was no
way to verify that this fixes the error.

REF: EKIRJASTO-68