Fix e-book reader
The turning of the screen made the late init fragments to show up as
null, so when they wer called, the reader would crash.
Added tags for each of the fragments, that can be used to look the
fragments up in the supportFragmentManager in onCreate and set them to
the correct variables. If there is no fragment with a matching tag in
the FragmentManager, we create a new one.