Make audiobookplayer turnable
All of the buttons and controls would pile on top of one another when
the player was turned.
Made the player turnable by making the view scrollable.

Adding the scrollbar broke the functionality of the seekbar, so changed
the use of OnTouchListener into using OnSeekbarChangeListener that was
recommended by the SeekBar.