Fix pdf reader
The pdf reader would crash when turning the device due to a late init
webview not being created.
Since the webview is always destroyed in onDestroy, which is done on
screen turn, we remove the savedInstance == null check, and always
recreate the the webvieFix pdf reader

The pdf reader would crash when turning the device due to a late init
webview not being created.
Since the webview is always destroyed in onDestroy, which is done on
screen turn, we remove the savedInstance == null check, and always
recreate the the webview.