Fix ebook crash and AndroidManifest changes
Added back the null check for reader2activity, as turning the screen
caused a fragment to be added twice to the fragment stack, which is
not alloed and caused a crash. Now it's only done when launched the
first time.

Also gave the app freedom to handle screenturning config changes by
itself, so now screen turning in audio book doesn't recreate the
activity and the experience is better for users. The same is done for
the main activity, which affects the magazines, and now turning screen
maintains the read spot.

Added styling to MAin activity for smoother look for the turned screen.

NOTE: This causes a old splash screen to pop up, with old palace logo,
that han be deleted from android.theme repo, but now there is double
splashscreens. This should be changed.