Remove maxSdkVersion definition from a permission
In the palace-audiobook library, FOREGROUND_SERVICE_MEDIA_PLAYBACK
permission was defined having maxSkdVersion 34, so the permission was
not requested for version 35, causing the audiobook to throw an error.

This is fixed by adding the same permission into the app's
AndroidManifest and removing the maxSdkVersion, so the permission is
asked even on newer versions.

REF: EKIRJASTO-194