Add automatic refresh to magazines,to book actions
Make the automatic refresh implementation to books and magazines, so
that if 401 HTTP error is noticed, the app tries to refresh the token
immediately.
If this is successful, the user sees a slightly longer
loading screen or a screen  refresh.
If unsuccessful, the user is informed about expired session and is then
shown the login page.
Most of the refreshes  are triggered and handled by the controller, but
the refresh is possible to be triggered from other places as well.
Fir instance, the magazines fragment and the dependents view trigger the
refresh on their own. This is because the controller has all the needed
information but it only handles book related actions, which magazines
and dependents are not.
But the code was put into controller, as it kept the fragments as
oblivious to the background happenings as possible, and it was more
simple to put into the controller, that is shared, rather than multiple
different viewModels.
If logout happens due to expired token, and the token can not be
refreshed, the user is logged out without actually deleting the user's
data, as deleting the data would only cause irritation if the logout is
not spesifically triggered by the user.

The way the dependents fragment was coded proved difficult to make
asynchronous so that the refresh triggering would work correctly, so
the viewmodel is changed to match the magazine fragment more, and thus
there is new files that handle the http requests and the fragment now
has a state. The state could also be used to handle the error messaging
and other things seen in the fragment, but kept the old inplementation
of that as it would've taken more time to ensure the visiblilities of
all the different parts were correct.

Also deleted the refresh button from the settings page.

REF: EKIRJASTO-110