Add special chapter skip behavior for speed change
There was a problem with player having empty space at the end of the
player when playing the audiobook on faster speeds.
This might be thanks to deprecated code, but this pr provides some sort
of solution that is done locally, without needing changes to the
palace-audiobook library. Note that updating the player to some higher
version could fix this.

This solution does some computations to count the player soit that the
player should be on if the timing was working correctly.
The audiobook length is adjusted based on the playerspeed, but the
smaller times on the edges only move faster, not reflecting the actual
time.

The idea is that we manully compute when the player should skip to the
next chapter and then do it manually. The way this is done is by keeping
count of what is the last spot we jump to in the player, and as the
player doesn't recognize (or at least doensn't update us accordingly)
what moment the player is currently on, and it thinks it's still in
normal time, leabcing empty space in the end of the file, we consider
the time onwards from the last spot to be quickened up, and oupate our
seek bar accordingly. Then when our seek bar reaches its end, we should
be in a spot where the audiobook chapter is done, and we manually skip
to the next chapter.