Add basic functions and icons for selected
Added a pickup of the possibly excisting selected Datetime into the OBDS
feed and parser, so it can be used to change what user sees. Added
handling for the selectedURI available in AuthenticationDocument.
Current returns the loans.
Added test button to load selected from settings page.
Added the select and unselect functions to the controller.
Added an icon that changes depending on if the user has the book
selected or not to the catalog views.

Needs to be added still:
Button for single book view
401 handling
Move seeing selected books to a different place
Add little toast to tell that a book has been added or deleted
Fix some error handling

Ref: EKIR-88