Ensure correct functioning of tasks and visuals
Fixed both the select and unselect tasks to handle the database and book
register correctly, as well as added comments to the code.
Added a toast that informs user about selecting and unselecting a book
being successful. Added functions that handle the toasts based on book
status.
Added functions to reset a book status after the toast has been showed.
Added accessibility strings and descriptions, though unsure if they are
used. Also added conditions to when a book is shown as revokes in the
accessibility section.

TODO:
- Change the location of select button in app + rename
- Reload selected on remove from the list
- Making into local feed
- Handling of 401
- Show selected only in selected and book info page