Replace faq link with user instructions
Replaces the faq link in the settings page with user instructions link.
Creates three language versions for the link, as nearly the full link is
different in the different language versions.

This version still uses old configureDocViewButton function, but it
doesn't replace the language "tag" like it still does for the other
links, and instead just creates the correct onClick listener.

This should be changed when all the other links are handled similarly to
these links.

TODO: Have other links function the same way aswell

Ref: EKIRJASTO-116