Move selected location and ensure correct funtions
Moved the selected feed to now be shown in the same bottom tab as loans.
Removing a loan removed the entry from database and register, so added a
selected check, so the book is only removed if it's also not selected.
Added translatable strings for the different tabs and ensured correct
functioning even when feed is empty (used to not show the nvigation tabs
if feed was empty).
Added descriptions to functions.
Removed the seected button from the settings page.
Ensured that foltering and sorting in the new views works correctly.

NOTE: All favorites are now stored in the local database which might
cause slowness and storage spage filling if there are alot of them.