Add the handling od 401 to selection functions
Made select and unselect tasks extend the AbstractBookTask class, so
that the 401 error can be caught and handled in the controller.
If select or unselect trigger token refresh and it's successful, the
view is reloaded, but the popup does still show, though it is delayed
due to the login behaviour.

SyncBook behaviour was also adjusted, as it was triggering the refresh
twice in a row, causing the first one cto succeed and the next one fail,
causing the user to be logged out unnecessarily.