Add thranslated titles to bottom bar buttons
The buttons were always in english, so add the correct titles to the
buttons that are translatable. Also fix the text for the sleep timer
to a shorter version.

Do these changes for both of the text files that are copies of one
anouther, just in case the files need to match.