Adjust the bottom bar and fix book returning
Moved the tabs around so that loans and reservation are under one tab,
and that selected books are on another, TO this second tab, new tabs can
be added for other lists.

Revoking a book worked incorrectly when a book was selected, causing an
old state of the book to be stored in registry. Fixed this.

There also was a stalement issue with the views on logging out and back
in. Fixed this by popping the backstack of the botton fragments, so that
they are recreated the next time the views are looked at.

Hid the collection buttons for all languages.