Fix sleep timer and bookmarks
Jumping to a bookmark from the same chapter, from a point that was
earlier than the bookmark caused the player to jump to next chapter.
Fixed this by adding a check to see if the player made jumps that we
were unaware of. Bookmarks were also placed incorrectly on sped up
books, so fixed that by first jumping to current spot so when we call
the player to make a bookmark from current spot, it is the spot the user
is hearing.

The end to chapter sleep timer did not work, so made a check to our new
end of chapter check to see if we have sleep timer with end of chapter
running, and pausing if we do. Also changed the initial value of sleep
timer, so instead of always starting with end of chapter timer running,
we have no timer running and make end of chapter timer if need be.

When the book ends, the book now pauses itself on all speeds. There is
code for a popup that asks if the user wants to return the book, but
that is never triggered. This could be fixed in later audiobook player
versions.