Disable sync bookmark toggle button functionality
All Kotlin code used to handle the sync bookmarks toggle button
in Settings tab of the E-kirjasto app is now commented out.
Note that no code was removed in this task.

The toggle was only visible to the logged-in user,
and if enabled, the user's book bookmarks should have been
synchronised between the application and the server.
However, since this feature is not available in E-kirjasto,
we are disabling the functions and variables used to display
the sync toggle button in the UI for now.

Ref. EKIRJASTO-287, EKIRJASTO-288