EKIRJASTO-287 Hide bookmark sync feature in UI (#193)
* Hide sync bookmark toggle button in UI

All XML code used to display the bookmark sync toggle
in Settings tab of E-kirjasto app is now commented out.
Note that no code was not removed in this task.

The toggle was only visible to the logged-in user,
and if enabled, the user's book bookmarks should have been
synchronised between the application and the server.
However, since this feature is not available in E-kirjasto,
we are hiding the toggle and related texts from the UI for now.

Ref. EKIRJASTO-287, EKIRJASTO-288

* Disable sync bookmark toggle button functionality

All Kotlin code used to handle the sync bookmarks toggle button
in Settings tab of the E-kirjasto app is now commented out.
Note that no code was removed in this task.

The toggle was only visible to the logged-in user,
and if enabled, the user's book bookmarks should have been
synchronised between the application and the server.
However, since this feature is not available in E-kirjasto,
we are disabling the functions and variables used to display
the sync toggle button in the UI for now.

Ref. EKIRJASTO-287, EKIRJASTO-288

* Add  comments to clarify the code being commented-out

Ref. EKIRJASTO-287, EKIRJASTO-288