Fix skip back behaviour when jumpin to start of book (#201)
After last changes, pressing the jump back button when less than 30 seconds of the book had been played caused the pointer to jump to the beginning, but not the audiobook itself. This was due to the AudioBookPlayer not jumping back if there was no actual point to jump to.

This is now fixed by computing how much of the book we have listened, and if it's less than 30 seconds, we jump to the beginning of the book. Speeding up the book causes the jump to beginning happen seemingly from further away than 30 seconds, but only a few seconds further away. It shouldn't be too noticeable for users.

NOTE: If the jump to the beginning happens from another chapter than the first one, the audiobook pauses.