Rework the notification permission popup
Added a new information popup that happens if the user has denied the popup permission before, that explains what the permissions are used for. Like before, the asking of the permission happens on launch. Now, after accepting or denying the permission, there is a toast that informs of the choice the user has made.