Add null check to refresh request
If for some reason, when refresh request returns from the server, and out login is in a state where there are no credentials, (like logged out or login failed or such states) the app would crash due to expected credentials value.

Added a null check to the credentials, and if there are none, instead of trying to refresh the token, mark the magazines load  as failed, and continue on normally.